로그인 관련 자바 스크립트 모듈화 및 불필요한 코드 제거
