+++ Index.html
@@ -48,8 +48,8 @@
       </div>
     </header>
     <main>
-      <div class="content1">
-        <div class="container1">
+      <div class="headContent">
+        <div class="headContainer">
           <h2>
             <span>세상의 모든 정보</span>를 <br/>
             <data-lineChanger>쉽게 저장하고</data-lineChanger>
