+++ linkbrary.css
@@ -46,7 +46,7 @@ a {
   line-height: normal;
 }
 
-.content1 {
+.headContent {
   display: flex;
   margin-top: 93px;
   margin-bottom: 70px;
@@ -58,7 +58,7 @@ a {
   overflow: hidden;
 }
 
-.container1 {
+.headContainer {
   display: flex;
   flex-direction: column;
   gap: 40px;
@@ -67,7 +67,7 @@ a {
   max-height: 833px;
 }
 
-.container1 > h2 {
+.headContainer > h2 {
   color: #000;
   font-family: Pretendard;
   font-size: 64px;
@@ -75,7 +75,7 @@ a {
   line-height: 125%;
 }
 
-.container1 > h2 > span {
+.headContainer > h2 > span {
   color: transparent;
   background: linear-gradient(91deg, #6d6afe 17.28%, #ff9f9f 74.98%);
   background-clip: text;
@@ -222,13 +222,13 @@ footer {
     max-width: 800px;
   }
 
-  .content1 {
+  .headContent {
     margin-top: 93px;
     margin-bottom: 30px;
     padding: 40px 251px 0px;
   }
 
-  .container1 {
+  .headContainer {
     max-height: 716px;
   }
 
@@ -282,23 +282,23 @@ footer {
     font-size: 14px;
   }
 
-  .content1 {
+  .headContent {
     display: flex;
     margin-top: 62px;
     padding: 28px 32px 0px;
   }
 
-  .container1 {
+  .headContainer {
     gap: 24px;
     max-height: 481px;
   }
 
-  .container1 > h2 {
+  .headContainer > h2 {
     font-size: 32px;
-    line-height: 42px;
+    line-height: 1.3125;
   }
 
-  .container1 > h2 > span {
+  .headContainer > h2 > span {
     color: transparent;
     background: linear-gradient(91deg, #6d6afe 17.28%, #ff9f9f 74.98%);
     background-clip: text;
