+++ signin.html
@@ -58,7 +58,7 @@
               비밀번호를 확인해 주세요.
             </div>
           </div>
-          <button class="signbutton">로그인</button>
+          <button type="submit" class="signbutton">로그인</button>
         </form>
       </div>
       <div class="socialsign">
@@ -83,6 +83,6 @@
         </div>
       </div>
     </main>
-    <script src="signin.js"></script>
+    <script type="module" src="signin.js"></script>
   </body>
 </html>
